Disable doctests
Starting with `random-1.3.0` doctests are executed in a CI by invoking
`doctest` executable directly, which turns out to be the correct way to
use doctest. This was pointed out in #165

In order to avoid `random-1.2.x` major version depending on doctest, this
patch removes doctest usage completely, in favor of a new setup going
forward. However, in order to prevent tooling breakaage downstream that
runs tests we do not remove the `doctest` executable section.